+++ mitra_activitypub/src/lib.rs
@@ -15,4 +15,5 @@ pub mod identity;
 pub mod importers;
 mod ownership;
 pub mod queues;
+pub mod utils;
 pub mod vocabulary;